+++ scripts/terrain_elevation.js
@@ -1,7 +1,9 @@
 /* globals
 canvas,
 game,
-_levels
+CONFIG,
+ui,
+PIXI
 */
 "use strict";
 
@@ -10,17 +12,33 @@ Used by ruler to get elevation at waypoints and at the end of the ruler.
 */
 
 import { log, MODULE_ID } from "./module.js";
+import { SETTINGS, getSetting } from "./settings.js";
 
 /**
  * Retrieve the elevation at the current ruler origin.
- * This is either token elevation or terrain elevation or 0.
+ * This is either the measuring token elevation or terrain elevation or 0.
+ * Cached during a ruler movement
  */
 export function elevationAtOrigin() {
   const measuringToken = this._getMovementToken();
   const origin = this.waypoints[0];
-  return measuringToken
-    ? tokenElevation(measuringToken)
-    : this.terrainElevationAtPoint(origin, { considerTokens: false });
+  if ( !origin ) return undefined;
+  if ( origin._terrainElevation ) return origin._terrainElevation;
+
+  let value = 0;
+
+  // Use the measuring token's elevation, if any, as the starting point.
+  if ( measuringToken ) value = tokenElevation(measuringToken);
+
+  // If the Levels UI is enabled, start at the bottom of the current layer.
+  else if ( useLevels() && CONFIG.Levels.UI.rangeEnabled ) value = parseFloat(CONFIG.Levels.UI.range[0]);
+
+  // Otherwise, get the elevation for this origin point.
+  else value = this.terrainElevationAtPoint(origin, { considerTokens: false });
+
+  origin._terrainElevation = value;
+  origin._userElevationIncrements = 0;
+  return value;
 }
 
 /**
@@ -31,17 +49,27 @@ export function elevationAtOrigin() {
  * @returns {number}
  */
 function tokenElevation(token) {
-  return token.document?.elevation ?? 0;
+  return token?.document?.elevation ?? 0;
+}
+
+/**
+ * Determine if token elevation should be preferred
+ * @returns {boolean}
+ */
+function preferTokenElevation() {
+  if ( !getSetting(SETTINGS.PREFER_TOKEN_ELEVATION) ) return false;
+  const token_controls = ui.controls.controls.find(elem => elem.name === "token");
+  const prefer_token_control = token_controls.tools.find(elem => elem.name === SETTINGS.PREFER_TOKEN_ELEVATION);
+  return prefer_token_control.active;
 }
 
 /**
  * Retrieve the terrain elevation at the current ruler destination
+ * @param {object} [options]  Options that modify the calculation
+ * @param {boolean} [options.considerTokens]    Consider token elevations at that point.
  */
-export function terrainElevationAtDestination({
-  startingElevation = 0,
-  considerTokens = true,
-  ignoreBelow = Number.NEGATIVE_INFINITY } = {}) {
-  return this.terrainElevationAtPoint(this.destination, { startingElevation, considerTokens, ignoreBelow });
+export function terrainElevationAtDestination({ considerTokens = true } = {}) {
+  return this.terrainElevationAtPoint(this.destination, { considerTokens });
 }
 
 /**
@@ -54,39 +82,40 @@ export function terrainElevationAtDestination({
  *
  * @param {Point} p      Point to measure, in {x, y} format
  * @param {object} [options]  Options that modify the calculation
- * @param {number} [options.startingElevation]  The elevation of the ruler or measuring token, used for Levels
  * @param {boolean} [options.considerTokens]    Consider token elevations at that point.
- * @param {number} [options.ignoreBelow]        If elevation of a given source is below this number, ignore
  * @returns {number} Elevation for the given point.
  */
-export function terrainElevationAtPoint({x, y}, {
-  startingElevation = 0,
-  considerTokens = true,
-  ignoreBelow = Number.NEGATIVE_INFINITY } = {}) {
-  log(`Checking Elevation at (${x}, ${y}) ${(considerTokens ? "" : "not ") + "considering tokens"}; ignoring below ${ignoreBelow}`);
+export function terrainElevationAtPoint(p, { considerTokens = true } = {}) {
+
+  const measuringToken = this._getMovementToken();
+  const startingElevation = this.elevationAtOrigin();
+  const ignoreBelow = ( measuringToken && preferTokenElevation() ) ? startingElevation : Number.NEGATIVE_INFINITY;
+
+  log(`Checking Elevation at (${p.x}, ${p.y}) ${(considerTokens ? "" : "not ") + "considering tokens"}\n\tstarting elevation ${startingElevation}\n\tignoring below ${ignoreBelow}`);
 
   if ( considerTokens ) {  // Check for tokens; take the highest one at a given position
     const tokens = retrieveVisibleTokens();
     const max_token_elevation = tokens.reduce((e, t) => {
       // Is the point within the token control area?
-      if ( !t.hitArea.contains(x, y) ) return e;
-      return Math.max(t.data.elevation, e);
+      if ( !t.bounds.contains(p.x, p.y) ) return e;
+      return Math.max(tokenElevation(t), e);
     }, Number.NEGATIVE_INFINITY);
-    log(`calculateEndElevation: ${tokens.length} tokens at (${x}, ${y}) with maximum elevation ${max_token_elevation}`);
+    log(`calculateEndElevation: ${tokens.length} tokens at (${p.x}, ${p.y}) with maximum elevation ${max_token_elevation}`);
 
     // Use tokens rather than elevation if available
     if ( isFinite(max_token_elevation) && max_token_elevation >= ignoreBelow ) return max_token_elevation;
   }
 
-  const ev_elevation = EVElevationAtPoint({x, y});
-  if ( levels_elevation !== undefined && levels_elevation > ignoreBelow ) return ev_elevation;
-
-  // Try levels
-  const levels_elevation = LevelsElevationAtPoint({x, y}, startingElevation);
+  // Try Levels
+  const levels_elevation = LevelsElevationAtPoint(p, { startingElevation });
   if ( levels_elevation !== undefined && levels_elevation > ignoreBelow ) return levels_elevation;
 
-  // Try terrain
-  const terrain_elevation = TerrainLayerElevationAtPoint({x, y});
+  // Try Elevated Vision
+  const ev_elevation = EVElevationAtPoint(p);
+  if ( ev_elevation !== undefined && ev_elevation > ignoreBelow ) return ev_elevation;
+
+  // Try Enhanced Terrain Layer
+  const terrain_elevation = TerrainLayerElevationAtPoint(p);
   if ( terrain_elevation !== undefined && terrain_elevation > ignoreBelow ) return terrain_elevation;
 
   // Default to 0 elevation for the point
@@ -97,6 +126,34 @@ function retrieveVisibleTokens() {
   return canvas.tokens.children[0].children.filter(c => c.visible);
 }
 
+// ----- HELPERS TO TRIGGER ELEVATION MEASURES ---- //
+/**
+ * Should Elevated Vision module be used?
+ * @returns {boolean}
+ */
+function useElevatedVision() {
+  return game.modules.get("elevatedvision")?.active
+    && game.settings.get(MODULE_ID, "enable-elevated-vision-elevation");
+}
+
+/**
+ * Should Terrain Layers module be used?
+ * @returns {boolean}
+ */
+function useTerrainLayer() {
+  return game.modules.get("enhanced-terrain-layer")?.active
+    && game.settings.get(MODULE_ID, "enable-enhanced-terrain-elevation");
+}
+
+/**
+ * Should Levels module be used?
+ * @returns {boolean}
+ */
+function useLevels() {
+  return game.modules.get("levels")?.active
+    && game.settings.get(MODULE_ID, "enable-levels-elevation");
+}
+
 // ----- ELEVATED VISION ELEVATION ----- //
 /**
  * Measure the terrain elevation at a given point using Elevated Vision.
@@ -104,10 +161,9 @@ function retrieveVisibleTokens() {
  * @returns {Number|undefined} Point elevation or undefined if elevated vision layer is inactive
  */
 function EVElevationAtPoint({x, y}) {
-  if ( !game.modules.get("elevatedvision")?.active ) return undefined;
+  if ( !useElevatedVision() ) return undefined;
   return canvas.elevation.elevationAt(x, y);
 }
-
 
 // ----- TERRAIN LAYER ELEVATION ----- //
 /**
@@ -117,8 +173,7 @@ function EVElevationAtPoint({x, y}) {
  * @return {Number|undefined} Point elevation or undefined if terrain layer is inactive or no terrain found.
  */
 function TerrainLayerElevationAtPoint({x, y}) {
-  if ( !game.settings.get(MODULE_ID, "enable-terrain-elevation")
-    || !game.modules.get("enhanced-terrain-layer")?.active ) return undefined;
+  if ( !useTerrainLayer() ) return undefined;
 
   const terrains = canvas.terrain.terrainFromPixels(x, y);
   if ( terrains.length === 0 ) return undefined; // No terrains found at the point.
@@ -143,13 +198,10 @@ function TerrainLayerElevationAtPoint({x, y}) {
 // use cases:
 // generally:
 // - if over a level-enabled object, use the bottom of that level.
-// - if multiple, use the bottom
-// - if hole, use the bottom
+// - if multiple, use the closest to the starting elevation
 // starting point of the ruler is a token:
 // - if the same level is present, stay at that level
 //   (elevation should be found from the token, so no issue)
-// - if a hole, go to bottom of the hole
-// - display level as labeled in the levels object flag?
 
 /*
  * Measure the elevation of any levels tiles at the point.
@@ -158,65 +210,36 @@ function TerrainLayerElevationAtPoint({x, y}) {
  * @param {PIXI.Point} p    Point to measure, in {x, y} format.
  * @return {Number|undefined} Levels elevation or undefined if levels is inactive or no levels found.
  */
-function LevelsElevationAtPoint(p, starting_elevation) {
-  if ( !game.settings.get(MODULE_ID, "enable-levels-elevation") || !game.modules.get("levels")?.active ) {
-    return undefined;
-  }
+function LevelsElevationAtPoint(p, { startingElevation = 0 } = {}) {
+  if ( !useLevels() ) return undefined;
 
-  // If in a hole, use that
-  const hole_elevation = checkForHole(p, starting_elevation);
-  if ( hole_elevation !== undefined ) return hole_elevation;
+  let tiles = [...levelsTilesAtPoint(p)];
+  if ( !tiles.length ) return undefined;
 
-  // Use levels if found
-  const levels_objects = _levels.getFloorsForPoint(p); // @returns {Object[]} returns an array of object each containing {tile,range,poly}
-  log("LevelsElevationAtPoint levels_objects", levels_objects);
-  return checkForLevel(p, starting_elevation);
+  tiles = tiles
+    .filter(t => startingElevation >= t.document.flags.levels.rangeBottom && startingElevation < t.document.flags.levels.rangeTop)
+    .sort((a, b) => a.document.flags.levels.rangeBottom - b.document.flags.levels.rangeBottom);
+
+  const ln = tiles.length;
+  if ( !ln ) return undefined;
+  return tiles[ln - 1].document.flags.levels.rangeBottom;
 }
 
-// function levelNameAtPoint(p, zz) {
-//   if ( !game.settings.get(MODULE_ID, "enable-levels-elevation") || !game.modules.get("levels")?.active ) {
-//     return undefined;
-//   }
-//
-//   const floors = _levels.getFloorsForPoint(p);
-//   if ( !floors || floors.length < 1 ) { return undefined; }
-//
-//   const levels_data = canvas.scene.getFlag("levels", "sceneLevels"); // Array with [0]: bottom; [1]: top; [2]: name
-//   if ( !levels_data ) { return undefined; }
-//   for ( let l of levels_data ) {
-//     if ( zz <= l[1] && zz >= l[0] ) return l[2];
-//   }
-//   return undefined;
-// }
 
+/**
+ * Get all tiles that have a levels range
+ * @param {Point} {x, y}
+ * @returns {Set<Tile>}
+ */
+function levelsTilesAtPoint({x, y}) {
+  const bounds = new PIXI.Rectangle(x, y, 1, 1);
+  const collisionTest = (o, rect) => { // eslint-disable-line no-unused-vars
+    // The object o constains n (Quadtree node), r (rect), t (object to test)
+    const flags = o.t.document?.flags?.levels;
+    if ( !flags ) return false;
+    if ( !isFinite(flags.rangeTop) || !isFinite(flags.rangeBottom) ) return false;
+    return true;
+  };
 
-// Check for level; return bottom elevation
-function checkForLevel(intersectionPT, zz) {
-  // Poly undefined for tiles.
-  const floors = _levels.getFloorsForPoint(intersectionPT); // @returns {Object[]} returns an array of object each containing {tile,range,poly}
-  log("checkForLevel floors", floors);
-  const floor_range = findCurrentFloorForElevation(zz, floors);
-  log(`checkForLevel current floor range for elevation ${zz}: ${floor_range[0]} ${floor_range[1]}`);
-  if ( !floor_range ) return undefined;
-  return floor_range[0];
-}
-
-function findCurrentFloorForElevation(elevation, floors) {
-  for ( let floor of floors ) {
-  if ( elevation <= floor.range[1] && elevation >= floor.range[0] )
-    return floor.range;
-  }
-  return false;
-}
-
-// Check if a floor is hollowed by a hole
-// Based on Levels function, modified to return bottom elevation of the hole.
-function checkForHole(intersectionPT, zz) {
-  for ( let hole of _levels.levelsHoles ) {
-    const hbottom = hole.range[0];
-    const htop = hole.range[1];
-    if ( zz > htop || zz < hbottom ) continue;
-    if ( hole.poly.contains(intersectionPT.x, intersectionPT.y) ) return hbottom;
-  }
-  return undefined;
+  return canvas.tiles.quadtree.getObjects(bounds, { collisionTest });
 }